requirement.txt
streamlit
scikit-learn
pandas
matplotlib
wordcloud
pillow
